# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BettingAI — Assistant IA de Paris Sportifs
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  scheduled: 'scheduled',
  live: 'live',
  finished: 'finished',
} as const;

export interface Match {
  id: number;
  homeTeam: string;
  awayTeam: string;
  league: string;
  country: string;
  kickoffAt: string;
  status: MatchStatus;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  topPrediction: string;
  topOdd?: number;
  confidence: number;
  isValueBet?: boolean;
}

export type MatchDetailStatus = typeof MatchDetailStatus[keyof typeof MatchDetailStatus];


export const MatchDetailStatus = {
  scheduled: 'scheduled',
  live: 'live',
  finished: 'finished',
} as const;

export interface Prediction {
  market: string;
  marketLabel?: string;
  selection: string;
  selectionLabel?: string;
  confidence: number;
  reasoning: string;
  oddValue: number;
  isValueBet?: boolean;
}

export interface OddEntry {
  bookmaker: string;
  market: string;
  selection: string;
  oddValue: number;
}

export interface MatchDetail {
  id: number;
  homeTeam: string;
  awayTeam: string;
  league: string;
  country: string;
  kickoffAt: string;
  status: MatchDetailStatus;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  homeForm: string;
  awayForm: string;
  predictions: Prediction[];
  odds?: OddEntry[];
}

export interface CouponRequest {
  targetOdd: number;
  /** @nullable */
  bookmaker?: string | null;
}

export interface CouponSelection {
  matchId: number;
  homeTeam: string;
  awayTeam: string;
  market: string;
  marketLabel: string;
  selection: string;
  selectionLabel: string;
  oddValue: number;
  confidence: number;
  reasoning?: string;
}

export type CouponStatus = typeof CouponStatus[keyof typeof CouponStatus];


export const CouponStatus = {
  pending: 'pending',
  won: 'won',
  lost: 'lost',
} as const;

export type CouponLevel = typeof CouponLevel[keyof typeof CouponLevel];


export const CouponLevel = {
  prudent: 'prudent',
  standard: 'standard',
  risque: 'risque',
} as const;

export interface Coupon {
  id: number;
  targetOdd: number;
  actualOdd: number;
  selections: CouponSelection[];
  confidenceAvg: number;
  bookmaker: string;
  status: CouponStatus;
  level: CouponLevel;
  /** @nullable */
  warning?: string | null;
  createdAt: string;
}

export interface CouponVariants {
  prudent: Coupon;
  standard: Coupon;
  risque: Coupon;
}

export interface RecentResult {
  date: string;
  market: string;
  selection: string;
  correct: boolean;
  confidence: number;
}

export interface PerformanceStats {
  period: string;
  totalPredictions: number;
  correctPredictions: number;
  winRate: number;
  avgConfidence: number;
  bestMarket: string;
  worstMarket: string;
  recentResults: RecentResult[];
}

export interface DashboardSummary {
  totalMatchesToday: number;
  liveMatchesCount: number;
  scheduledMatchesCount: number;
  finishedMatchesCount: number;
  totalCouponsToday: number;
  avgConfidence: number;
  valueBetsCount: number;
  topLeague: string;
}

export interface League {
  id: number;
  name: string;
  country: string;
  matchCount: number;
  logo: string;
}

export type GetTodaysMatchesParams = {
league?: string;
status?: GetTodaysMatchesStatus;
};

export type GetTodaysMatchesStatus = typeof GetTodaysMatchesStatus[keyof typeof GetTodaysMatchesStatus];


export const GetTodaysMatchesStatus = {
  scheduled: 'scheduled',
  live: 'live',
  finished: 'finished',
} as const;

export type GetAiPerformanceParams = {
period?: GetAiPerformancePeriod;
};

export type GetAiPerformancePeriod = typeof GetAiPerformancePeriod[keyof typeof GetAiPerformancePeriod];


export const GetAiPerformancePeriod = {
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
  all: 'all',
} as const;