# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BettingAI — Assistant IA de Paris Sportifs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/today:
    get:
      operationId: getTodaysMatches
      tags: [matches]
      summary: Get all matches of the day with AI predictions
      parameters:
        - name: league
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [scheduled, live, finished]
      responses:
        "200":
          description: List of today's matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a single match with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Match not found
  /matches/{id}/analysis:
    get:
      operationId: getMatchAnalysis
      tags: [matches]
      summary: Get AI analysis for a specific match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
        "404":
          description: Match not found
...
components:
  schemas:
...
    Match:
      type: object
      required: [id, homeTeam, awayTeam, league, country, kickoffAt, status, topPrediction, confidence]
      properties:
        id:
          type: integer
        homeTeam:
          type: string
        awayTeam:
          type: string
        league:
          type: string
        country:
          type: string
        kickoffAt:
          type: string
        status:
          type: string
          enum: [scheduled, live, finished]
        homeScore:
          type: ["integer", "null"]
        awayScore:
          type: ["integer", "null"]
        topPrediction:
          type: string
        topOdd:
          type: number
        confidence:
          type: number
        isValueBet:
          type: boolean
    MatchDetail:
      type: object
      required: [id, homeTeam, awayTeam, league, country, kickoffAt, status, homeForm, awayForm, predictions]
      properties:
        id:
          type: integer
        homeTeam:
          type: string
        awayTeam:
          type: string
        league:
          type: string
        country:
          type: string
        kickoffAt:
          type: string
        status:
          type: string
          enum: [scheduled, live, finished]
        homeScore:
          type: ["integer", "null"]
        awayScore:
          type: ["integer", "null"]
        homeForm:
          type: string
        awayForm:
          type: string
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/Prediction"
        odds:
          type: array
          items:
            $ref: "#/components/schemas/OddEntry"
    Prediction:
      type: object
      required: [market, selection, confidence, reasoning, oddValue]
      properties:
        market:
          type: string
        marketLabel:
          type: string
        selection:
          type: string
        selectionLabel:
          type: string
        confidence:
          type: number
        reasoning:
          type: string
        oddValue:
          type: number
        isValueBet:
          type: boolean
    OddEntry:
      type: object
      required: [bookmaker, market, selection, oddValue]
      properties:
        bookmaker:
          type: string
        market:
          type: string
        selection:
          type: string
        oddValue:
          type: number
...