# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BettingAI — Assistant IA de Paris Sportifs
 * OpenAPI spec version: 0.1.0
 */
import type { MatchDetailStatus } from './matchDetailStatus';
import type { OddEntry } from './oddEntry';
import type { Prediction } from './prediction';

export interface MatchDetail {
  id: number;
  homeTeam: string;
  awayTeam: string;
  league: string;
  country: string;
  kickoffAt: string;
  status: MatchDetailStatus;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  homeForm: string;
  awayForm: string;
  predictions: Prediction[];
  odds?: OddEntry[];
}