# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BettingAI — Assistant IA de Paris Sportifs
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  market: string;
  marketLabel?: string;
  selection: string;
  selectionLabel?: string;
  confidence: number;
  reasoning: string;
  oddValue: number;
  isValueBet?: boolean;
}